feat: Add method to load synchronized lyrics into state
The code changes introduce a new method in the `SyncedLyricsNotifier` class to load synchronized lyrics into the state. This method takes a track name, artist name, and lyrics as input and parses the lyrics to store them as a list of maps in the state. The commit message follows the established convention of using a prefix to indicate the type of change, in this case, "feat" for a new feature.